create tf_util for terraform helper functions
