+++ pkg/tools/opal/opal.go
@@ -10,6 +10,7 @@ import (
 	"github.com/soluble-ai/soluble-cli/pkg/download"
 	"github.com/soluble-ai/soluble-cli/pkg/log"
 	"github.com/soluble-ai/soluble-cli/pkg/tools"
+	tf_util "github.com/soluble-ai/soluble-cli/pkg/tools/util"
 	"github.com/soluble-ai/soluble-cli/pkg/util"
 	"github.com/spf13/cobra"
 )
@@ -85,7 +86,7 @@ func (t *Tool) Run() (*tools.Result, error) {
 	}
 
 	if t.EnableModuleDownload {
-		err := t.runTerraformGet()
+		err := tf_util.RunTerraformGet(t.GetDirectory(), t.RunOpts)
 		if err != nil {
 			log.Warnf("{warning:terraform get} failed ")
 			result.AddValue("TERRAFORM_GET_FAILED", "true")
